% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_structure_proportions.R
\name{age_structure_proportions}
\alias{age_structure_proportions}
\title{Determine the proportions of the age structure of the population}
\usage{
age_structure_proportions(path = mypath, runs = 1, gen = 49, species = 0)
}
\arguments{
\item{path}{path to the simulation folder}

\item{runs}{integer referring to the number of montecarlo simulations output files to examine. It refers to the mcruns option in the RunVars CDMetaPOP input file. Right now it defaults to 1, which would be only one montecarlo run.}

\item{gen}{Simulation run time (generation or year). Refers to runtime in the RunVars of CDMetaPOP input file. Currently defaults to 49.}

\item{species}{If the simulation includes more species... The code for this needs to be adjusted}
}
\value{
a dataframe with a  column for each montecarlo run (MC) with the proportion of the population for each given age (in rows) and the last column with the average value across all montecarlo runs at any given age.
}
\description{
This function summarizes the proportions of the age structure in the simulated population. It is calculated counting the values in the column 'age' of the chosen 'ind' generation.
}
\examples{
define  your path to the output_test folder: mypath = "~/UM/CDMetaPop/EBT/EBT_Inputs/data/output_test1728487700/"
}
