% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersal.R
\name{dispersal}
\alias{dispersal}
\title{This function counts the number of individuals that move each year by reading in each ind file}
\usage{
dispersal(
  path = path,
  run = 0,
  batch = 0,
  mc = 0,
  gen = 9,
  species = 0,
  ind = TRUE
)
}
\arguments{
\item{path}{path to the simulation folder}

\item{run}{A run is each row of the RunVars file - Fix description}

\item{mc}{integer referring to the number of montecarlo simulations output files to examine. It refers to the mcruns option in the RunVars CDMetaPOP input file. Right now it defaults to 0, which would be only one  run.}

\item{gen}{Simulation run time (generation or year). Refers to runtime in the RunVars of CDMetaPOP input file. Currently defaults to 49}

\item{species}{If the simulation includes more species... The code for this needs to be adjusted. Right now, species = 0, means one species only and it is the default}

\item{ind}{If ind is true, the function will return calculations based on the ind files, otherwise on the sample files}
}
\value{
a dataframe The data returned is a dataframe that lists the proportion of dispersors for each year and for each mc simulation
}
\description{
This function counts the number of individuals that move each year by reading in each ind file
}
\examples{
define  your path to the output_test folder: mypath = "your_path/your_output_test/" For example "~/UM/CDMetaPop/EBT/EBT_Inputs/data/output_test1728487700/"
}
