% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic_diagnostics.R
\name{hets_plot}
\alias{hets_plot}
\title{Create heterozygosity plot}
\usage{
hets_plot(data)
}
\arguments{
\item{x}{dataframe summary_popAllTime.csv}
}
\value{
a ggplot object representing the overtime observed and expected heterozygosities values in the simulation.
}
\description{
This function plots heterozygosity overtime. It is based on the Ho and He columns of the CDMetaPOP output file summary_popAllTime
}
