% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_size_overtime.R
\name{myy_progeny_plot}
\alias{myy_progeny_plot}
\title{Plot the proportion of exclusive Myy progeny over the total}
\usage{
myy_progeny_plot(file_path)
}
\arguments{
\item{x}{Specify the path to the summary_popAllTime.csv dataframe}
}
\value{
a ggplot object representing the overtime progeny size of YYMales
}
\description{
This function plots the ratio of the Myy progeny overtime
}
