% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_size_overtime.R
\name{pop_mature_plot}
\alias{pop_mature_plot}
\title{Create abundance by year plot showing the numbers of mature individuals of all sexes}
\usage{
pop_mature_plot(file_path)
}
\arguments{
\item{x}{Specify the path to the summary_popAllTime.csv dataframe}
}
\value{
a ggplot object representing the overtime population size of sexually mature individuals
}
\description{
This function plots population dynamics overtime showing a count of the mature individuals over the years of the simulation separated by their sex. It is based off the N_Mature columns of the CDMetaPOP output file summary_popAllTime
}
