% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_size_overtime.R
\name{pop_year_class}
\alias{pop_year_class}
\title{Create abundance by year for each age class}
\usage{
pop_year_class(file_path, n = 5)
}
\arguments{
\item{n}{An integer that specifies the time intervals of the time series. It defaults to 5 years.}

\item{x}{Specify the path to the summary_classAllTime.csv dataframe}
}
\value{
a ggplot object representing the  general population size at 5 years intervals for the different age classes.
}
\description{
This function plots population dynamics overtime according by the age of the individuals. It shows the count of the individuals over the years of the simulation separated by age. It is based off the N_Initial_Class column of the CDMetaPOP output file summary_classAllTime
}
