% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_conversions.R
\name{separate_column}
\alias{separate_column}
\title{Column symbol separator}
\usage{
separate_column(file_path, column_name, sep)
}
\arguments{
\item{x}{Specify the file name and path.}
}
\value{
Dataframe of columns for each value that was originally separated by the delimiter.
}
\description{
This function eases the separation of column content generated from cdmetapop outputs. CDMetaPOP uses delimiters and different delimiter options are specified for different fields. Delimiter ‘|’ or ‘bar’ is used time parameters. The delimiter ‘~’ (‘tilda’) is used to assign parameters based on sex. The delimiter ‘;’ is the default and used in various contexts to split fields and patches parameters. The delimiter ‘:’ is used to split parameters used for functions. Note that Loo uses one instance of ‘;’ in an Hindex option. See CDMetaPOP manual for more specific uses and examples.
}
