% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_conversions.R
\name{unite_column}
\alias{unite_column}
\title{Unite multiple columns into one}
\usage{
unite_column(dataframe, column_name, sep)
}
\arguments{
\item{x}{A data frame}
}
\value{
Description of the returned value.
}
\description{
This replicates the function 'unite' from tidyr which can actually be used :)
}
